requirements.txt
streamlit>=1.29.0
pandas>=2.0.0
plotly>=5.14.0
requests>=2.28.2
python-dotenv>=1.0.0
psycopg2-binary>=2.9.5  # For direct DB access when needed